+++ Requests/Bookings/Create.php
@@ -14,7 +14,7 @@ class Create extends AbstractRequest
     /**
      * @inheritdoc
      */
-    protected $endpoint = '/bookings';
+    protected $endpoint = '/api/bookings';
 
     /**
      * Constructor.