+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2019-03-26
+
+**Added**
+
+* `Bookings\Confirm` request to confirm bookings on invoice.
+* `Config\Fetch` request to request the API's configuration.
+
+**Changed**
+
+* Move the `/api` suffix from the hostname towards the request endpoints.
+
+
 ## [0.5.0] - 2019-03-15
 
 **Added**